fix: keeping policy data consistent. deInit doesn't seem to exist
